# docs/public/api/swagger-2.0.yaml
basePath: /api
definitions:
...
  repo.Group:
    properties:
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  repo.GroupInvitation:
    properties:
      expiresAt:
        type: string
      group:
        $ref: '#/definitions/repo.Group'
      id:
        type: string
      uses:
        type: integer
    type: object
...
  repo.GroupUpdate:
    properties:
      currency:
        type: string
      name:
        type: string
    type: object
...
  repo.PaginationResult-repo_ItemSummary:
    properties:
      items:
        items:
          $ref: '#/definitions/repo.ItemSummary'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  v1.ActionAmountResult:
    properties:
      completed:
        type: integer
    type: object
...
  v1.CreateRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  v1.ItemAttachmentToken:
    properties:
      token:
        type: string
    type: object
...
  v1.WipeInventoryOptions:
    properties:
      wipeLocations:
        type: boolean
      wipeMaintenance:
        type: boolean
      wipeTags:
        type: boolean
    type: object
...
host: demo.homebox.software
info:
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  version: "1.0"
paths:
  /v1/actions/create-missing-thumbnails:
    post:
      description: Creates thumbnails for items that are missing them
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Create Missing Thumbnails
      tags:
      - Actions
  /v1/actions/ensure-asset-ids:
    post:
      description: Ensures all items in the database have an asset ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Ensure Asset IDs
      tags:
      - Actions
  /v1/actions/ensure-import-refs:
    post:
      description: Ensures all items in the database have an import ref
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Ensures Import Refs
      tags:
      - Actions
  /v1/actions/set-primary-photos:
    post:
      description: Sets the first photo of each item as the primary photo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Set Primary Photos
      tags:
      - Actions
  /v1/actions/wipe-inventory:
    post:
      description: Deletes all items in the inventory
      parameters:
      - description: Wipe options
        in: body
        name: options
        schema:
          $ref: '#/definitions/v1.WipeInventoryOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Wipe Inventory
      tags:
      - Actions
  /v1/actions/zero-item-time-fields:
    post:
      description: Resets all item date fields to the beginning of the day
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Zero Out Time Fields
      tags:
      - Actions
  /v1/assets/{id}:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.PaginationResult-repo_ItemSummary'
      security:
      - Bearer: []
      summary: Get Item by Asset ID
      tags:
      - Items
...
  /v1/groups:
    delete:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Group
      tags:
      - Group
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Group'
      security:
      - Bearer: []
      summary: Get Group
      tags:
      - Group
    post:
      parameters:
      - description: Create group request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.Group'
      security:
      - Bearer: []
      summary: Create Group
      tags:
      - Group
    put:
      parameters:
      - description: User Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.GroupUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Group'
      security:
      - Bearer: []
      summary: Update Group
      tags:
      - Group
  /v1/groups/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.Group'
            type: array
      security:
      - Bearer: []
      summary: Get All Groups
      tags:
      - Group
  /v1/groups/invitations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.GroupInvitation'
            type: array
      security:
      - Bearer: []
      summary: Get All Group Invitations
      tags:
      - Group
    post:
      parameters:
      - description: User Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.GroupInvitationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GroupInvitation'
      security:
      - Bearer: []
      summary: Create Group Invitation
      tags:
      - Group
  /v1/groups/invitations/{id}:
    delete:
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Group Invitation
      tags:
      - Group
    post:
      parameters:
      - description: Invitation Token
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GroupAcceptInvitationResponse'
      security:
      - Bearer: []
      summary: Accept Group Invitation
      tags:
      - Group
  /v1/groups/members:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.UserSummary'
            type: array
      security:
      - Bearer: []
      summary: Get All Group Members
      tags:
      - Group
    post:
      parameters:
      - description: User ID
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.GroupMemberAdd'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Add User to Group
      tags:
      - Group
  /v1/groups/members/{user_id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Remove User from Group
      tags:
      - Group
  /v1/groups/statistics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.GroupStatistics'
      security:
      - Bearer: []
      summary: Get Group Statistics
      tags:
      - Statistics
  /v1/groups/statistics/locations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.TotalsByOrganizer'
            type: array
      security:
      - Bearer: []
      summary: Get Location Statistics
      tags:
      - Statistics
  /v1/groups/statistics/purchase-price:
    get:
      parameters:
      - description: start date
        in: query
        name: start
        type: string
      - description: end date
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ValueOverTime'
      security:
      - Bearer: []
      summary: Get Purchase Price Statistics
      tags:
      - Statistics
  /v1/groups/statistics/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.TotalsByOrganizer'
            type: array
      security:
      - Bearer: []
      summary: Get Tags Statistics
      tags:
      - Statistics
  /v1/items:
    get:
      parameters:
      - description: search string
        in: query
        name: q
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: items per page
        in: query
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: tags Ids
        in: query
        items:
          type: string
        name: tags
        type: array
      - collectionFormat: multi
        description: location Ids
        in: query
        items:
          type: string
        name: locations
        type: array
      - collectionFormat: multi
        description: parent Ids
        in: query
        items:
          type: string
        name: parentIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.PaginationResult-repo_ItemSummary'
      security:
      - Bearer: []
      summary: Query All Items
      tags:
      - Items
    post:
      parameters:
      - description: Item Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.ItemSummary'
      security:
      - Bearer: []
      summary: Create Item
      tags:
      - Items
  /v1/items/{id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Item
      tags:
      - Items
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Get Item
      tags:
      - Items
    patch:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Item Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - Items
    put:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Item Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - Items
  /v1/items/{id}/attachments:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: File attachment
        in: formData
        name: file
        required: true
        type: file
      - description: Type of file
        in: formData
        name: type
        type: string
      - description: Is this the primary attachment
        in: formData
        name: primary
        type: boolean
      - description: name of the file including extension
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
      security:
      - Bearer: []
      summary: Create Item Attachment
      tags:
      - Items Attachments
  /v1/items/{id}/attachments/{attachment_id}:
    delete:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Item Attachment
      tags:
      - Items Attachments
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ItemAttachmentToken'
      security:
      - Bearer: []
      summary: Get Item Attachment
      tags:
      - Items Attachments
    put:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      - description: Attachment Update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.ItemAttachmentUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Update Item Attachment
      tags:
      - Items Attachments
  /v1/items/{id}/duplicate:
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Duplicate Options
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.DuplicateOptions'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.ItemOut'
      security:
      - Bearer: []
      summary: Duplicate Item
      tags:
      - Items
  /v1/items/{id}/maintenance:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - enum:
        - scheduled
        - completed
        - both
        in: query
        name: status
        type: string
        x-enum-varnames:
        - MaintenanceFilterStatusScheduled
        - MaintenanceFilterStatusCompleted
        - MaintenanceFilterStatusBoth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.MaintenanceEntryWithDetails'
            type: array
      security:
      - Bearer: []
      summary: Get Maintenance Log
      tags:
      - Item Maintenance
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Entry Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.MaintenanceEntryCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.MaintenanceEntry'
      security:
      - Bearer: []
      summary: Create Maintenance Entry
      tags:
      - Item Maintenance
  /v1/items/{id}/path:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.ItemPath'
            type: array
      security:
      - Bearer: []
      summary: Get the full path of an item
      tags:
      - Items
  /v1/items/export:
    get:
      responses:
        "200":
          description: text/csv
          schema:
            type: string
      security:
      - Bearer: []
      summary: Export Items
      tags:
      - Items
  /v1/items/fields:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Names
      tags:
      - Items
  /v1/items/fields/values:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - Bearer: []
      summary: Get All Custom Field Values
      tags:
      - Items
  /v1/items/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image to upload
        in: formData
        name: csv
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Import Items
      tags:
      - Items
...
  /v1/labelmaker/item/{id}:
    get:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Print this label, defaults to false
        in: query
        name: print
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: image/png
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get Item label
      tags:
      - Items
  /v1/labelmaker/location/{id}:
    get:
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Print this label, defaults to false
        in: query
        name: print
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: image/png
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get Location label
      tags:
      - Locations
  /v1/locations:
    get:
      parameters:
      - description: Filter locations with parents
        in: query
        name: filterChildren
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.LocationOutCount'
            type: array
      security:
      - Bearer: []
      summary: Get All Locations
      tags:
      - Locations
    post:
      parameters:
      - description: Location Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.LocationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.LocationSummary'
      security:
      - Bearer: []
      summary: Create Location
      tags:
      - Locations
  /v1/locations/{id}:
    delete:
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete Location
      tags:
      - Locations
    get:
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.LocationOut'
      security:
      - Bearer: []
      summary: Get Location
      tags:
      - Locations
    put:
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Location Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/repo.LocationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.LocationOut'
      security:
      - Bearer: []
      summary: Update Location
      tags:
      - Locations
...
  /v1/notifiers/test:
    post:
      parameters:
      - description: URL
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Test Notifier
      tags:
      - Notifiers
...
schemes:
- https
- http
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"